refactor: Extract api_fetch into separate module and require auth
Move api_fetch logic from cli.rs to fetch.rs for better organization.
Add authentication check to require login before making API requests.
