+++ src/cli.rs
@@ -10,6 +10,7 @@ use crate::auth;
 use crate::config::{self, Config};
 use crate::context::CommandContext;
 use crate::feature;
+use crate::fetch::api_fetch;
 #[cfg(feature = "feedback")]
 use crate::feedback::{self, FeedbackType};
 use crate::help;
@@ -481,46 +482,6 @@ fn get_subcommand(name: &str) -> clap::Command {
         .expect("subcommand should exist")
 }
 
-async fn api_fetch(
-    ctx: &mut CommandContext,
-    method: &str,
-    url: &str,
-    query_args: Option<&str>,
-    data: Option<&str>,
-    json: Option<&str>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let method_upper = method.to_uppercase();
-    let mut request = match method_upper.as_str() {
-        "GET" => ctx.client.get(url),
-        "POST" => ctx.client.post(url),
-        "PUT" => ctx.client.put(url),
-        "PATCH" => ctx.client.patch(url),
-        "DELETE" => ctx.client.delete(url),
-        _ => return Err(format!("Unsupported HTTP method: {}", method).into()),
-    };
-    if let Some(args) = query_args {
-        let pairs: Vec<(&str, &str)> = args
-            .split(',')
-            .filter_map(|pair| pair.split_once('='))
-            .collect();
-        request = request.query(&pairs);
-    }
-    if let Some(body) = data {
-        request = request.body(body.to_string());
-    }
-    if let Some(body) = json {
-        let parsed: serde_json::Value =
-            serde_json::from_str(body).map_err(|e| format!("Invalid JSON: {}", e))?;
-        request = request.json(&parsed);
-    }
-    let response = request.send().await?;
-    let status = response.status();
-    let body = response.text().await?;
-    println!("{}", status);
-    println!("{}", body);
-    Ok(())
-}
-
 #[derive(Parser)]
 #[command(
     name = "ana",
@@ -820,10 +781,4 @@ mod tests {
         );
     }
 
-    #[tokio::test]
-    async fn test_api_healthcheck() {
-        let mut ctx = CommandContext::new();
-        let result = api_fetch(&mut ctx, "GET", "/api/auth/healthcheck", None, None, None).await;
-        assert!(result.is_ok());
-    }
 }
