+++ src/main.rs
@@ -5,6 +5,7 @@ mod config;
 mod context;
 mod diagnostics;
 mod feature;
+mod fetch;
 #[cfg(feature = "feedback")]
 mod feedback;
 mod help;
